Create first factories and IIFE
For the moment I don't really know where I'm going and so what I need
to create. But the basics are the gameboard as an IIFE because I only
need one, the Player factory and Game factory. I have to figure out how
to transcribe a tic-tac-toe game on an array.