Create the console game with much pain
I don't know why but I didn't want to draft a plan of the programm so
it was hard to know where to start. But I managed to build the game only
through the console. I've added the condition for the position to be
empty, and the win condition (big bulky if). ot proud but not motivated.