Finish the function showing the gameboard
I also added the winnerState to show when a round is won. It's just the
background in red. I have to figure out how to freeze the buttons so you
can't keep playing after someone has won. I think I have to add a state
one the game factory to control the start and end of a round (and so I
have to rename the playRound into play).