Create inRound state and functions to start end
I found a way to disable the gameboard when a round is won/tied, by
creating a state of the round and to give a condition on the creation
of the eventlisteners to be in a round. So now when someone won/tied,
it displays the gameboar but without the eventlisteners. I now have to
build some other html elements, to start/restart a round and later to
create the players and display their scores.