Built first on page elements like player forms
I want the page to load on the player choosing name state so you have
to enter a name for both players and then create the game and start the
first round. The button to restart a round will appear once the game is
created (for now I let it be here so I don't forget its structure).