Update the scores afetr a win
I went wayyyy over my head with trying to ceate a game object when the
players are ready and by clicking an initial button... But there's no
way to use an object created inside an evenlistener function ! So I went
easier by keeping my playerOne and playerTwo and game objects, the only
3 lines in the global scope. Then I removed the creation of the score
inside of the form eventlisteners... such an idiot. And created a simple
refreshPlayersScores function to call after a win.